requirements.txt
openpyxl==3.1.5
pandas==3.0.2
pillow==12.2.0
streamlit==1.56.0
gspread==6.2.1
google-auth==2.49.2
pytest==9.0.3
